refactor: move InstanceIndexUpdateService to hale.ui plugin
Moves the implementation of InstanceIndexUpdateService from its own separate
plugin to the common `hale.ui` plugin. The previously existing ServiceFactory
is dropped and creation of the service is moved to `HaleServiceFactory`
instead.

This is done to avoid a circular dependency between the `hale.ui` and the
removed `hale.ui.index` plugin that would be introduced when trying to
use `InstanceIndexUpdateService` in the `AbstractInstanceService` class in
the future. This is desirable to synchronize recreation of the instance
index and running a retransformation after an alignment change.
