+++ ui/features/eu.esdihumboldt.hale.ui.feature/feature.xml
@@ -490,11 +490,4 @@
          install-size="0"
          version="0.0.0"/>
 
-   <plugin
-         id="eu.esdihumboldt.hale.ui.index"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
 </feature>
