--- ui/plugins/eu.esdihumboldt.hale.ui.index/.settings/org.eclipse.core.resources.prefs
@@ -1,4 +1,0 @@
-#Created from default preferences 01.12.2017 14:26:09
-#Fri Dec 01 14:26:09 CET 2017
-eclipse.preferences.version=1
-encoding/<project>=UTF-8
