--- ui/plugins/eu.esdihumboldt.hale.ui.index/.settings/org.eclipse.jdt.groovy.core.prefs
@@ -1,4 +1,0 @@
-#Created from default preferences 28 Oct 2022, 08:10:30
-#Fri Oct 28 08:10:30 CEST 2022
-eclipse.preferences.version=1
-groovy.compiler.level=25
