--- ui/plugins/eu.esdihumboldt.hale.ui.index/META-INF/MANIFEST.MF
@@ -1,22 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Instance index
-Bundle-SymbolicName: eu.esdihumboldt.hale.ui.index;singleton:=true
-Bundle-Version: 5.1.0.qualifier
-Bundle-Vendor: wetransform GmbH
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Import-Package: com.google.common.collect;version="17.0.0",
- eu.esdihumboldt.cst.functions.groovy,
- eu.esdihumboldt.hale.common.instance.index,
- eu.esdihumboldt.hale.ui.service.align,
- eu.esdihumboldt.hale.ui.service.instance
-Require-Bundle: eu.esdihumboldt.hale.common.core;bundle-version="3.4.0",
- eu.esdihumboldt.hale.common.align;bundle-version="3.4.0",
- eu.esdihumboldt.hale.common.instance;bundle-version="3.4.0",
- org.eclipse.ui;bundle-version="3.107.0",
- org.eclipse.core.runtime;bundle-version="3.11.1",
- eu.esdihumboldt.hale.ui;bundle-version="3.4.0",
- eu.esdihumboldt.hale.common.schema;bundle-version="3.4.0",
- eu.esdihumboldt.hale.common.instance.groovy,
- eu.esdihumboldt.util.groovy;bundle-version="3.4.0"
-Automatic-Module-Name: eu.esdihumboldt.hale.ui.index
