--- ui/plugins/eu.esdihumboldt.hale.ui.index/build.properties
@@ -1,5 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .,\
-               plugin.xml
