+++ ui/plugins/eu.esdihumboldt.hale.ui/META-INF/MANIFEST.MF
@@ -69,6 +69,7 @@ Import-Package: com.google.common.base;version="1.6.0",
  eu.esdihumboldt.hale.common.instance.geometry,
  eu.esdihumboldt.hale.common.instance.geometry.impl,
  eu.esdihumboldt.hale.common.instance.helper,
+ eu.esdihumboldt.hale.common.instance.index,
  eu.esdihumboldt.hale.common.instance.io,
  eu.esdihumboldt.hale.common.instance.model,
  eu.esdihumboldt.hale.common.instance.model.impl,
