+++ ui/plugins/eu.esdihumboldt.hale.ui/plugin.xml
@@ -645,6 +645,9 @@
          <service
                serviceClass="eu.esdihumboldt.hale.common.tasks.TaskService">
          </service>
+         <service
+               serviceClass="eu.esdihumboldt.hale.ui.service.index.InstanceIndexUpdateService">
+         </service>
       </serviceFactory>
       <sourceProvider
             provider="eu.esdihumboldt.hale.ui.service.schema.internal.SchemaServiceSource">
